refactor vite configuration
